InferenceScripts/requirements.txt
faster-whisper
openai-whisper
sagemaker
boto3
alt-profanity-check